Ignore unit tests
Make all the wallet unit tests into a single test and use `ignore`. This is
because it is unlikely that this test will be run often, it makes changes to the
filesystem and requires tight control on how monero-wallet-rpc was started.

The test is still useful to prove [once off] that this code works.