Just make it work.
The monero-wallet needs to stay up to sync with monerod. This takes some time. I've added some polling logic which verifies if the wallet is at the same height as the node, however, even if, it looks like that their might be race conditions where the wallet does not see the funds :S