Implement check_tx_key
The wallet RPC call `check_tx_key` is used to get information about a
transaction on the blockchain using the `tx_key`. The `tx_key` is the so called
'view key' i.e., a private key created for each transaction that allows one to
view the transaction details.

Add requisite flags to `transfer` as required by `monero-wallec-rpc` so we get
back the `tx_key` when doing a transfer. Implement `check_tx_key`.